% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LoadR}
\alias{LoadR}
\title{LoadR'}
\usage{
LoadR(DT)
}
\arguments{
\item{DT}{This is a data.table object}
}
\value{
a data.table with all the data
}
\description{
This function 'loads' the data in the right format (data.table)
}
\examples{
DT = data.table::data.table(plyr::baseball)
DT = LoadR(DT)
}
