% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ScrubR}
\alias{ScrubR}
\title{ScrubR'}
\usage{
ScrubR(DT, predictors, rowWiseMissingPercentage, SDs, transformTable,
  scalingMethod, imputationMethod)
}
\arguments{
\item{DT}{This is a data.table object}

\item{predictors}{These are the clean predictors provided manually or from the output of ScanR}

\item{rowWiseMissingPercentage}{This the the allowed threshold of the percentage missing data 
per row}

\item{SDs}{This the the allowed threshold of standard deviations away from the mean}

\item{transformTable}{This is the table received from ScanR that provides information what 
features and how they need to be transformed}

\item{scalingMethod}{This is the way the data will be scaled}

\item{imputationMethod}{This is the way missing data will be treated}
}
\value{
a data.table with all the preprocessed data
}
\description{
This function 'scrubs' the data, it massages the data in checking and fixing missing data,
transformations, scaling and outlier removal or imputation
}
\examples{
DT = data.table::data.table(plyr::baseball)
DT = LoadR(DT)
rs = scanr(DT, correlationCutoff = 0.99, SamplingPercentage=15, percentUniqueCutoff=5)
if (length(names(DT)) == length(unique(names(DT)))) {
data.table::setnames(DT, names(DT), rs$Predictors)
} else {
	names(DT) = rs$Predictors
}
predictors = rs[Status=='Healthy' | Status=='Uniform' ,Predictors]
transformTable = rs[abs(Skewness)>abs(selectedTransformationBoundary),
c('Predictors', 'Skewness'), with=F]
results = scrubr(DT, predictors, rowWiseMissingPercentage=100, SDs=8,
transformTable=transformTable, scalingMethod='robustZscore', imputationMethod='CWD')
}
