% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SmashR}
\alias{SmashR}
\title{SmashR'}
\usage{
SmashR(DT, UoA, predictors)
}
\arguments{
\item{DT}{This is a data.table object}

\item{UoA}{This is the variable in the data.table describing the unit of analysis}

\item{predictors}{These are the clean predictors provided manually or from the output of ScanR}
}
\value{
a data.table with all the statistics per UoA
}
\description{
This function 'smashes' the data, it calculates various statistics per unit of analysis (UoA)
}
\examples{
DT = data.table::data.table(plyr::baseball)
UoA = 'team'
predictors = c('g', 'ab', 'r', 'h')
rs = SmashR(DT, UoA, predictors)
}
